+++ ARCH.md
@@ -13,10 +13,6 @@ src-->otherImpl...
 include-->sys
 include-->arch_h
 arch_h-->generic
-system-->bare
-system-->rtthread
-system-->common
-system-->otherSys..
 arch-->arm
 arch-->otherArch...
 ```
@@ -45,38 +41,24 @@ include目录中有三种类型的文件
 
 ### 3. arch
 
-存放不同架构对于一些硬件相关接口的实现
+存放不同架构对于一些硬件/系统相关接口的实现
 
 各个架构的硬件接口实现各占用一个文件夹
-
-### 4. system
-
-system目录中有三种类型的文件
-
-- common：存放不同的系统的通用
-- bare：裸机状态下的资源管理代码
-- othersystem：在带有操作系统状态下的资源管理代码
 
 ## 系统调用接口
 
 ### 接口的设计思路
 
-众所周知，接口就是我们用来获取系统资源的通道，而在mlibc中系统调用接口分为两种：**裸机接口和操作系统接口**。
+众所周知，接口就是我们用来获取系统资源的通道，mlibc通过这些接口来对上层提供服务。
 
-裸机接口可能看着比较陌生，它其实就是将裸机状态下的一些资源抽象为系统资源了，我们仍然通过系统调用的方式去使用它，但实际它并没有系统管理。通过这种方式，我们不仅可以将硬件资源掌握在自己手里（按需修改资源管理代码），同时也能享受上层c库提供的一些方便好用的函数。
-
-目前，我们只提供了两种类别的接口：**文件IO接口和内存管理接口**。
+接口默认是裸机实现，在不同系统/硬件的环境中通过重载这些接口来提供服务。
 
 ```mermaid
 flowchart TD
-user([user])
-user--调用-->mlibc
-mlibc((mlibc))
-mlibc--依赖-->文件IO
-mlibc--依赖-->内存管理
+mlibc--依赖-->lib_nosys
+硬件/系统--重载-->lib_nosys
+lib_nosys--实现-->interface
 ```
-
-
 
 **文件IO接口提供的服务有：**
 
@@ -85,13 +67,11 @@ mlibc--依赖-->内存管理
 3. 读文件
 4. 写文件
 5. 文件控制
+6. ......
 
 **内存管理接口提供的服务有：**
 
-1. 内存池初始化
-2. 内存申请
-3. 内存重申请
-4. 内存释放
+1. 堆区空间申请
 
 > 这里大家可能会有**疑问**：为什么裸机也要提供文件IO的服务呢？
 >
