+++ arch/SConscript
@@ -8,6 +8,9 @@ list = os.listdir(cwd)
 src     = []
 CPPPATH = [cwd]
 
+group = DefineGroup('mlibc', src, depend = [''], CPPPATH = CPPPATH)
+
+objs += group
 for d in list:
     path = os.path.join(cwd, d)
     if os.path.isfile(os.path.join(path, 'SConscript')):
