+++ src/internal/mem_impl.h
@@ -1,0 +1,6 @@
+#ifndef MLIBC_MEM_IMPL_H__
+#define MLIBC_MEM_IMPL_H__
+
+void __mlibc_sys_heap_init(void);
+
+#endif /* MLIBC_MEM_IMPL_H__*/
