--- src/posix/SConscript
@@ -1,10 +1,0 @@
-from building import *
-import os
-
-cwd     = GetCurrentDir()
-src     = Glob('*.c')
-CPPPATH = [cwd]
-
-group = DefineGroup('mlibc', src, depend = [''], CPPPATH = CPPPATH)
-
-Return('group')
