--- src/posix/close.c
@@ -1,19 +1,0 @@
-/*
- * Copyright (c) mlibc & plct lab
- *
- * SPDX-License-Identifier: MIT
- *
- * Change Logs:
- * Date           Author       Notes
- * 2024/5/28   0Bitbiscuits the first version
- */
-
-#include <unistd.h>
-#include <compiler.h>
-
-mlibc_weak int close(int d)
-{
-    (void)d;
-
-    return -1;
-}
