+++ src/stdlib/SConscript
@@ -5,6 +5,6 @@ cwd     = GetCurrentDir()
 src     = Glob('*.c')
 CPPPATH = [cwd]
 
-group = DefineGroup('__mlibc_stdlib', src, depend = [''], CPPPATH = CPPPATH)
+group = DefineGroup('mlibc', src, depend = [''], CPPPATH = CPPPATH)
 
 Return('group')
