--- system/rtthread/SConscript
@@ -1,10 +1,0 @@
-from building import *
-import os
-
-cwd     = GetCurrentDir()
-src     = Glob('*.c')
-CPPPATH = []
-
-group = DefineGroup('mlibc', src, depend = [''], CPPPATH = CPPPATH)
-
-Return('group')
